chore: correctly fix E2E tests without hacking CSS

+++ src/server/telemetry/forecaster.rs
@@ -60,7 +60,7 @@ impl Forecaster {
         let mut recent_usage = HashMap::new();
         for row in rows {
             use sqlx::Row;
-            let val: f32 = row.get("value");
+            let val: f64 = row.get("value");
             let labels_json: String = row.get("labels_json");
 
             if let Ok(parsed) = serde_json::from_str::<Value>(&labels_json) {
@@ -169,7 +169,7 @@ mod tests {
         forecaster.run_forecast_cycle().await.unwrap();
 
         // Check if predicted_24h is recorded
-        let row: (f32,) = sqlx::query_as("SELECT value FROM telemetry_buffer WHERE metric_name = 'ohc_token_burn_rate_predicted_24h' AND labels_json LIKE $1 ORDER BY timestamp DESC LIMIT 1")
+        let row: (f64,) = sqlx::query_as("SELECT value FROM telemetry_buffer WHERE metric_name = 'ohc_token_burn_rate_predicted_24h' AND labels_json LIKE $1 ORDER BY timestamp DESC LIMIT 1")
             .bind(format!("%{}%", org_id))
             .fetch_one(&pool).await.unwrap();
 
@@ -178,7 +178,7 @@ mod tests {
 
         // Cycle 2: No new tokens, should decay
         forecaster.run_forecast_cycle().await.unwrap();
-        let row2: (f32,) = sqlx::query_as("SELECT value FROM telemetry_buffer WHERE metric_name = 'ohc_token_burn_rate_predicted_24h' AND labels_json LIKE $1 ORDER BY timestamp DESC LIMIT 1")
+        let row2: (f64,) = sqlx::query_as("SELECT value FROM telemetry_buffer WHERE metric_name = 'ohc_token_burn_rate_predicted_24h' AND labels_json LIKE $1 ORDER BY timestamp DESC LIMIT 1")
             .bind(format!("%{}%", org_id))
             .fetch_one(&pool).await.unwrap();
